requirements.txt
-e .
aws-cdk.aws_ec2
aws-cdk.aws_s3
aws-cdk.aws_iam
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_autoscaling
aws-cdk.aws_ssm
aws-cdk.aws_secretsmanager
aws-cdk.aws_rds
aws-cdk.aws_lambda
aws-cdk.aws_logs
aws-cdk.aws_events
aws-cdk.aws_events_targets
